requirements.txt
opencv-python
Pillow
pymysql
matplotlib
python-dotenv
certifi
ultralytics
huggingface_hub
